requirements.txt
Django==2.0.3
datetime
django-tinymce4-lite
stripe